C4:246/571 - prevent corrupting vaults by front running them

+++ src/interfaces/IPublicVault.sol
@@ -24,7 +24,7 @@ interface IPublicVault is IVaultImplementation {
 
   struct VaultData {
     uint256 yIntercept;
-    uint48 slope;
+    uint256 slope;
     uint40 last;
     uint64 currentEpoch;
     uint256 withdrawReserve;
